taking care about 32 bits sample size in PulseDevice module in record mode

+++ modules/pulse.cpp
@@ -241,7 +241,7 @@ void PulseDeviceRead::readFromPaSimple() {
         say("error: no stream set.");
         return;
     }
-    u_int8_t data[4096]; //4Ko
+    u_int16_t data[4096]; //4Ko
     int bytesRead;
 
     int error = 0;
